Move InstrumentationModule#extend() implementation back to tooling/internal
