+++ javaagent-extension-api/src/main/java/io/opentelemetry/javaagent/extension/config/MethodsConfigurationParser.java
@@ -13,7 +13,7 @@ import java.util.Set;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-public class MethodsConfigurationParser {
+public final class MethodsConfigurationParser {
 
   private static final Logger log = LoggerFactory.getLogger(MethodsConfigurationParser.class);
