+++ javaagent-extension-api/src/main/java/io/opentelemetry/javaagent/extension/instrumentation/InstrumentationModule.java
@@ -5,21 +5,15 @@
 
 package io.opentelemetry.javaagent.extension.instrumentation;
 
-import static io.opentelemetry.javaagent.extension.matcher.AgentElementMatchers.failSafe;
 import static java.util.Arrays.asList;
 import static net.bytebuddy.matcher.ElementMatchers.any;
-import static net.bytebuddy.matcher.ElementMatchers.isAnnotatedWith;
-import static net.bytebuddy.matcher.ElementMatchers.named;
-import static net.bytebuddy.matcher.ElementMatchers.not;
 
 import io.opentelemetry.instrumentation.api.config.Config;
 import io.opentelemetry.javaagent.extension.AgentExtensionTooling;
 import io.opentelemetry.javaagent.extension.log.TransformSafeLogger;
 import io.opentelemetry.javaagent.extension.muzzle.InstrumentationClassPredicate;
-import io.opentelemetry.javaagent.extension.muzzle.Mismatch;
 import io.opentelemetry.javaagent.extension.muzzle.ReferenceMatcher;
 import io.opentelemetry.javaagent.extension.spi.AgentExtension;
-import java.security.ProtectionDomain;
 import java.util.ArrayList;
 import java.util.Collections;
 import java.util.LinkedHashSet;
@@ -28,13 +22,7 @@ import java.util.Map;
 import java.util.Set;
 import java.util.function.Predicate;
 import net.bytebuddy.agent.builder.AgentBuilder;
-import net.bytebuddy.description.annotation.AnnotationSource;
-import net.bytebuddy.description.method.MethodDescription;
-import net.bytebuddy.description.type.TypeDescription;
 import net.bytebuddy.matcher.ElementMatcher;
-import net.bytebuddy.utility.JavaModule;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
 
 /**
  * Instrumentation module groups several connected {@link TypeInstrumentation}s together, sharing
@@ -50,17 +38,11 @@ import org.slf4j.LoggerFactory;
 public abstract class InstrumentationModule implements AgentExtension {
   private static final TransformSafeLogger log =
       TransformSafeLogger.getLogger(InstrumentationModule.class);
-  private static final Logger muzzleLog = LoggerFactory.getLogger("muzzleMatcher");
 
   private static final String[] EMPTY = new String[0];
 
-  // Added here instead of AgentInstaller's ignores because it's relatively
-  // expensive. https://github.com/DataDog/dd-trace-java/pull/1045
-  public static final ElementMatcher.Junction<AnnotationSource> NOT_DECORATOR_MATCHER =
-      not(isAnnotatedWith(named("javax.decorator.Decorator")));
-
   private final Set<String> instrumentationNames;
-  protected final boolean enabled;
+  final boolean enabled;
 
   /**
    * Creates an instrumentation module. Note that all implementations of {@link
@@ -113,58 +95,18 @@ public abstract class InstrumentationModule implements AgentExtension {
    * @param parentAgentBuilder AgentBuilder to base instrumentation config off of.
    * @return the original agentBuilder and this instrumentation
    */
+  @Override
   public final AgentBuilder extend(AgentBuilder parentAgentBuilder, AgentExtensionTooling tooling) {
     if (!enabled) {
       log.debug("Instrumentation {} is disabled", extensionName());
       return parentAgentBuilder;
     }
+    return InstrumentationExtensionImplementation.get().extend(this, parentAgentBuilder, tooling);
+  }
 
-    List<String> helperClassNames = getAllHelperClassNames();
-    List<String> helperResourceNames = asList(helperResourceNames());
-    List<TypeInstrumentation> typeInstrumentations = typeInstrumentations();
-    if (typeInstrumentations.isEmpty()) {
-      if (!helperClassNames.isEmpty() || !helperResourceNames.isEmpty()) {
-        log.warn(
-            "Helper classes and resources won't be injected if no types are instrumented: {}",
-            extensionName());
-      }
-
-      return parentAgentBuilder;
-    }
-
-    ElementMatcher.Junction<ClassLoader> moduleClassLoaderMatcher = classLoaderMatcher();
-    MuzzleMatcher muzzleMatcher = new MuzzleMatcher(tooling);
-    AgentBuilder.Transformer helperInjector =
-        tooling.createHelperInjector(helperClassNames, helperResourceNames);
-    InstrumentationContextProvider contextProvider =
-        tooling.createInstrumentationContextProvider(getMuzzleContextStoreClasses());
-
-    AgentBuilder agentBuilder = parentAgentBuilder;
-    for (TypeInstrumentation typeInstrumentation : typeInstrumentations) {
-      AgentBuilder.Identified.Extendable extendableAgentBuilder =
-          agentBuilder
-              .type(
-                  failSafe(
-                      typeInstrumentation.typeMatcher(),
-                      "Instrumentation type matcher unexpected exception: " + getClass().getName()),
-                  failSafe(
-                      moduleClassLoaderMatcher.and(typeInstrumentation.classLoaderOptimization()),
-                      "Instrumentation class loader matcher unexpected exception: "
-                          + getClass().getName()))
-              .and(NOT_DECORATOR_MATCHER)
-              .and(muzzleMatcher)
-              .transform(ConstantAdjuster.instance())
-              .transform(helperInjector);
-      extendableAgentBuilder = contextProvider.instrumentationTransformer(extendableAgentBuilder);
-      extendableAgentBuilder =
-          applyInstrumentationTransformers(
-              typeInstrumentation.transformers(), extendableAgentBuilder, tooling);
-      extendableAgentBuilder = contextProvider.additionalInstrumentation(extendableAgentBuilder);
-
-      agentBuilder = extendableAgentBuilder;
-    }
-
-    return agentBuilder;
+  @Override
+  public final String extensionName() {
+    return instrumentationNames.iterator().next();
   }
 
   /**
@@ -176,83 +118,6 @@ public abstract class InstrumentationModule implements AgentExtension {
     helperClassNames.addAll(asList(additionalHelperClassNames()));
     helperClassNames.addAll(asList(getMuzzleHelperClassNames()));
     return helperClassNames;
-  }
-
-  private AgentBuilder.Identified.Extendable applyInstrumentationTransformers(
-      Map<? extends ElementMatcher<? super MethodDescription>, String> transformers,
-      AgentBuilder.Identified.Extendable agentBuilder,
-      AgentExtensionTooling tooling) {
-    for (Map.Entry<? extends ElementMatcher<? super MethodDescription>, String> entry :
-        transformers.entrySet()) {
-      agentBuilder =
-          agentBuilder.transform(
-              new AgentBuilder.Transformer.ForAdvice()
-                  .include(
-                      tooling.classLoaders().bootstrapProxyClassLoader(),
-                      tooling.classLoaders().agentClassLoader())
-                  .withExceptionHandler(tooling.adviceExceptionHandler())
-                  .advice(entry.getKey(), entry.getValue()));
-    }
-    return agentBuilder;
-  }
-
-  /**
-   * A ByteBuddy matcher that decides whether this instrumentation should be applied. Calls
-   * generated {@link ReferenceMatcher}: if any mismatch with the passed {@code classLoader} is
-   * found this instrumentation is skipped.
-   */
-  private class MuzzleMatcher implements AgentBuilder.RawMatcher {
-    private final AgentExtensionTooling tooling;
-
-    private MuzzleMatcher(AgentExtensionTooling tooling) {
-      this.tooling = tooling;
-    }
-
-    @Override
-    public boolean matches(
-        TypeDescription typeDescription,
-        ClassLoader classLoader,
-        JavaModule module,
-        Class<?> classBeingRedefined,
-        ProtectionDomain protectionDomain) {
-      /* Optimization: calling getMuzzleReferenceMatcher() inside this method
-       * prevents unnecessary loading of muzzle references during agentBuilder
-       * setup.
-       */
-      ReferenceMatcher muzzle = getMuzzleReferenceMatcher();
-      if (muzzle != null) {
-        boolean isMatch = muzzle.matches(tooling, classLoader);
-
-        if (!isMatch) {
-          if (muzzleLog.isWarnEnabled()) {
-            muzzleLog.warn(
-                "Instrumentation skipped, mismatched references were found: {} -- {} on {}",
-                extensionName(),
-                InstrumentationModule.this.getClass().getName(),
-                classLoader);
-            List<Mismatch> mismatches = muzzle.getMismatchedReferenceSources(tooling, classLoader);
-            for (Mismatch mismatch : mismatches) {
-              muzzleLog.warn("-- {}", mismatch);
-            }
-          }
-        } else {
-          if (log.isDebugEnabled()) {
-            log.debug(
-                "Applying instrumentation: {} -- {} on {}",
-                extensionName(),
-                InstrumentationModule.this.getClass().getName(),
-                classLoader);
-          }
-        }
-
-        return isMatch;
-      }
-      return true;
-    }
-  }
-
-  public final String extensionName() {
-    return instrumentationNames.iterator().next();
   }
 
   /**
