+++ javaagent-extension-api/src/test/resources/META-INF/services/io.opentelemetry.javaagent.extension.instrumentation.InstrumentationExtensionImplementation
@@ -1,0 +1,1 @@
+io.opentelemetry.javaagent.extension.instrumentation.TestInstrumentationExtensionImplementation
