+++ javaagent-tooling/src/main/java/io/opentelemetry/javaagent/extension/instrumentation/ConstantAdjuster.java
