+++ javaagent-tooling/src/main/java/io/opentelemetry/javaagent/tooling/context/InstrumentationContextProvider.java
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-package io.opentelemetry.javaagent.extension.instrumentation;
+package io.opentelemetry.javaagent.tooling.context;
 
 import net.bytebuddy.agent.builder.AgentBuilder;
