+++ javaagent-tooling/src/main/java/io/opentelemetry/javaagent/tooling/context/NoopContextProvider.java
@@ -5,7 +5,6 @@
 
 package io.opentelemetry.javaagent.tooling.context;
 
-import io.opentelemetry.javaagent.extension.instrumentation.InstrumentationContextProvider;
 import net.bytebuddy.agent.builder.AgentBuilder.Identified.Extendable;
 
 public class NoopContextProvider implements InstrumentationContextProvider {
